Remushter/requirements.txt
streamlit==1.22.0
opencv-python-headless==4.7.0.72
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
Pillow==9.5.0
matplotlib==3.7.1
albumentations==1.3.0
tqdm==4.66.1
